feat: inline InputFile
